requirements.txt
python-telegram-bot==21.5
qrcode[pil]==8.2
pillow==10.4.0
pyzbar==0.1.9
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
gspread==6.0.0
oauth2client==4.1.3
schedule==1.2.1